Added image grid utility
Now the contents will be shown on the image grid of the main page. In the future the contents which will be shown on the image grid will be marked with types and they will be put on the page accordingly, for now this is onyl a proof of concept, meaning the style will drastically change also.